requirements.txt
black
flake8
ipython
isort
jupyterlab
loguru
matplotlib
mkdocs
notebook
numpy
pandas
pip
pytest
python-dotenv
scikit-learn
tqdm
typer
-e .
pyyaml
scipy
